backend/requirements.txt
flask
flask-cors
langchain-groq
langchain-core
requests
python-dotenv